20201118_4

+++ yss/inc/drv/uart/drv_st_uart_type_B.h
@@ -50,7 +50,7 @@ namespace drv
 		Stream *mStream;
 
 	public :
-		Uart(USART_TypeDef *peri, void (*clockFunc)(bool en), void (*nvicFunc)(bool en), Stream *txStream, Stream *rxStream, unsigned char txChannel, unsigned char rxChannel, unsigned short priority, unsigned long (*getClockFreq)(void));
+		Uart(USART_TypeDef *peri, void (*clockFunc)(bool en), void (*nvicFunc)(bool en), Stream *txStream, Stream *rxStream, unsigned char txChannel, unsigned char rxChannel, unsigned short priority, unsigned int (*getClockFreq)(void));
 		bool init(unsigned int baud, unsigned int receiveBufferSize);
 		void isr(void);
 		void push(char data);
